# 271 simplifier cache (per request)