… test/rpc-integration/test-simplify/response-with-logging.json