Term simplifications are cached in a `Map Term Term` while simplifying, and supplied from cache when already present.

### Measurement with Foundry proofs

| Test                                                     | 271-simplifier-cache time | master time | (271-simplifier-cache/master) time 
| ----------------------------------------------------------|---------------------------|-------------|------------------------------------
| SymbolicStorageTest.testFail_SymbolicStorage(uint256)    | 196.76                    | 413.14      | 0.4762550225105291                 
| SymbolicStorageTest.testFail_SymbolicStorage1(uint256)   | 241.19                    | 464.52      | 0.5192241453543442                 
| AssumeTest.testFail_assume_true(uint256,uint256)         | 79.84                     | 111.79      | 0.7141962608462296                 
| AssumeTest.test_assume_false(uint256,uint256)            | 83.05                     | 113.27      | 0.7332038492098526                 
| test_foundry_resume_proof                                | 106.25                    | 136.2       | 0.7801027900146843                 
| EmitContractTest.testExpectEmitCheckEmitter()            | 85.33                     | 96.06       | 0.888298979804289                  
| EmitContractTest.testExpectEmit()                        | 84.19                     | 93.3        | 0.9023579849946409                 
| AccountParamsTest.test_getNonce_unknownSymbolic(address) | 110.19                    | 116.45      | 0.9462430227565478                 
| AssumeTest.test_multi_assume(address,address)            | 134.65                    | 141.59      | 0.9509852390705559                 
| AssumeTest.test_assume_staticCall(bool)                  | 55.51                     | 57.88       | 0.959053213545266                  
| AccountParamsTest.test_GetNonce_true()                   | 46.71                     | 45.03       | 1.0373084610259826                 
| AccountParamsTest.testDealSymbolic(uint256)              | 57.53                     | 55.28       | 1.0407018813314037                 
| BlockParamsTest.testFee(uint256)                         | 49.1                      | 47.12       | 1.0420203735144313                 
| ExpectRevertTest.testFail_expectRevert_multipleReverts() | 112.96                    | 107.66      | 1.0492290544306149                 
| BMCLoopsTest.test_countdown_symbolic(uint256)            | 136.61                    | 129.83      | 1.0522221366402218                 
| BMCLoopsTest.test_countdown_concrete()                   | 36.89                     | 34.85       | 1.0585365853658537                 
| UintTypeTest.test_uint256(uint256)                       | 39.11                     | 36.84       | 1.0616178067318132                 
| AssertTest.test_assert_true()                            | 54.12                     | 50.84       | 1.064516129032258                  
| BlockParamsTest.testBlockNumber()                        | 34.91                     | 32.32       | 1.0801361386138613                 
| BytesTypeTest.test_bytes32(bytes32)                      | 39.61                     | 34.44       | 1.1501161440185832                 
| BytesTypeTest.test_bytes4(bytes4)                        | 47.73                     | 38.45       | 1.2413524057217162                 
| TOTAL                                                    | 1832.2399999999998        | 2356.86     | 0.7774072282613306