@jberthold 2023-11-08
### KEVM performance comparison 
(with fix for map internalisation)

| Test                                                                                           | 271-simplifier-cache time |        master time | (271-simplifier-cache/master) time |
|------------------------------------------------------------------------------------------------|---------------------------|--------------------|------------------------------------|
|  erc20/ds/transferFrom-failure-2-c-spec.k   |                     55.71 |             365.41 |                0.15245888180400097 |
|  benchmarks/ecrecover00-sigvalid-spec.k     |                     94.79 |             470.89 |                0.20129966658879994 |
|  mcd/vat-deny-diff-fail-rough-spec.k        |                     95.53 |             449.78 |                0.21239272533238474 |
|  functional/evm-int-simplifications-spec.k  |                    111.47 |             345.96 |                 0.3222048791767835 |
|  mcd/end-subuu-pass-spec.k                  |                    139.34 |             430.78 |                0.32345977064859094 |
|  erc20/ds/approve-success-spec.k            |                    133.23 |             393.31 |                 0.3387404337545447 |
|  erc20/ds/transfer-failure-1-c-spec.k       |                     53.53 |             154.22 |                 0.3471015432499027 |
|  benchmarks/storagevar03-spec.k             |                    126.16 |             360.28 |                 0.3501720883757078 |
|  mcd/flopper-cage-pass-spec.k               |                     65.89 |             173.98 |                0.37872169214852286 |
|  erc20/ds/transferFrom-failure-1-d-spec.k   |                     66.05 |             150.94 |                 0.4375910957996555 |
|  examples/erc20-spec.md                     |                    307.74 |             693.98 |                 0.4434421741260555 |
|  erc20/hkg/transferFrom-failure-2-spec.k    |                     73.16 |              158.3 |                0.46216045483259627 |
|  benchmarks/bytes00-spec.k                  |                     55.01 |              114.5 |                 0.4804366812227074 |
|  erc20/ds/transferFrom-success-1-spec.k     |                     99.27 |             199.06 |                 0.4986938611473927 |
|  benchmarks/keccak00-spec.k                 |                     47.94 |              94.37 |                 0.5080004238635158 |
|  benchmarks/address00-spec.k                |                     41.82 |              80.93 |                 0.5167428642036327 |
|  benchmarks/requires01-a0le0-spec.k         |                     43.12 |              82.61 |                  0.521970705725699 |
|  erc20/ds/balanceOf-spec.k                  |                     46.31 |              88.39 |                 0.5239280461590677 |
|  erc20/hkg/allowance-spec.k                 |                     43.88 |               83.6 |                 0.5248803827751197 |
|  benchmarks/storagevar02-overflow-spec.k    |                     45.42 |              86.49 |                 0.5251474158862297 |
|  erc20/ds/totalSupply-spec.k                |                     43.27 |              81.03 |                 0.5339997531778354 |
|  erc20/ds/transferFrom-failure-2-a-spec.k   |                     76.64 |             141.54 |                 0.5414723753002685 |
|  benchmarks/staticloop00-a0lt10-spec.k      |                     59.69 |             109.69 |                 0.5441699334488103 |
|  erc20/hkg/transferFrom-success-1-spec.k    |                     73.49 |             134.45 |                 0.5465972480476013 |
|  bihu/forwardToHotWallet-success-2-spec.k   |                    132.97 |             243.16 |                  0.546841585787136 |
|  erc20/hkg/balanceOf-spec.k                 |                     45.47 |               82.5 |                 0.5511515151515152 |
|  erc20/hkg/transfer-success-1-spec.k        |                     56.69 |             102.54 |                  0.552857421494051 |
|  mcd/dsvalue-read-pass-spec.k               |                    172.76 |             305.86 |                 0.5648335839926764 |
|  bihu/forwardToHotWallet-failure-2-spec.k   |                     69.17 |             122.25 |                 0.5658077709611452 |
|  erc20/ds/transfer-success-1-spec.k         |                     73.03 |             128.49 |                 0.5683710794614366 |
|  mcd/vat-move-diff-rough-spec.k             |                    184.95 |             323.88 |                  0.571044831419044 |
|  erc20/ds/transferFrom-failure-2-b-spec.k   |                     93.17 |             162.89 |                 0.5719810915341642 |
|  erc20/ds/transferFrom-failure-1-c-spec.k   |                    104.87 |             182.84 |                 0.5735615838984905 |
|  erc20/ds/transferFrom-failure-1-b-spec.k   |                    138.09 |             239.37 |                 0.5768893345030706 |
|  mcd/flipper-bids-pass-rough-spec.k         |                    406.41 |              697.4 |                    0.5827502150846 |
|  mcd/vat-addui-pass-spec.k                  |                    235.84 |             403.96 |                  0.583820180215863 |
|  erc20/ds/transfer-failure-1-b-spec.k       |                      94.8 |             159.87 |                 0.5929817977106399 |
|  bihu/forwardToHotWallet-failure-4-spec.k   |                    107.85 |             181.09 |                 0.5955602186757966 |
|  mcd/dsvalue-peek-pass-rough-spec.k         |                    161.66 |              270.9 |                 0.5967515688445921 |
|  mcd/vat-subui-pass-spec.k                  |                    280.31 |              469.7 |                 0.5967851820310837 |
|  mcd/dsvalue-read-pass-summarize-spec.k     |                    127.81 |             213.58 |                 0.5984174548178668 |
|  erc20/ds/transferFrom-failure-1-a-spec.k   |                     88.97 |             148.07 |                 0.6008644560005403 |
|  mcd/dstoken-approve-fail-rough-spec.k      |                    105.64 |             175.53 |                 0.6018344442545434 |
|  mcd/flipper-ttl-pass-spec.k                |                    127.68 |             209.18 |                 0.6103834018548618 |
|  mcd/dai-symbol-pass-spec.k                 |                    129.16 |              211.5 |                 0.6106855791962175 |
|  erc20/ds/transfer-success-2-spec.k         |                     68.37 |             111.72 |                 0.6119763694951665 |
|  benchmarks/encode-keccak00-spec.k          |                      52.3 |              84.81 |                 0.6166725621978539 |
|  mcd/flipper-tau-pass-spec.k                |                    120.32 |             194.13 |                 0.6197908617936434 |
|  erc20/hkg/approve-spec.k                   |                      46.5 |              74.27 |                 0.6260939814191464 |
|  benchmarks/staticarray00-spec.k            |                     45.35 |              72.26 |                  0.627594796567949 |
|  mcd/functional-spec.k                      |                    415.37 |             659.17 |                 0.6301409348119605 |
|  benchmarks/ecrecover00-siginvalid-spec.k   |                    101.68 |             161.35 |                 0.6301828323520298 |
|  mcd/vat-addui-fail-rough-spec.k            |                    149.63 |             230.87 |                 0.6481136570364274 |
|  mcd/vat-subui-fail-rough-spec.k            |                    148.32 |             227.22 |                 0.6527594401901241 |
|  benchmarks/encodepacked-keccak01-spec.k    |                     45.12 |              68.82 |                 0.6556233653007847 |
|  mcd/vat-subui-pass-rough-spec.k            |                     219.6 |             334.81 |                 0.6558943878617723 |
|  erc20/ds/allowance-spec.k                  |                      47.8 |              70.58 |                 0.6772456786625106 |
|  erc20/ds/approve-failure-spec.k            |                      48.1 |              70.39 |                 0.6833357010939054 |
|  mcd/cat-exhaustiveness-spec.k              |                    241.81 |             352.71 |                  0.685577386521505 |
|  erc20/hkg/transferFrom-success-2-spec.k    |                    124.46 |             180.98 |                 0.6877002983755112 |
|  benchmarks/storagevar02-nooverflow-spec.k  |                     46.14 |              67.09 |                 0.6877328961097033 |
|  mcd/vat-slip-pass-rough-spec.k             |                    265.83 |             386.51 |                 0.6877700447595146 |
|  benchmarks/overflow00-nooverflow-spec.k    |                     42.73 |              61.93 |                 0.6899725496528338 |
|  erc20/hkg/transferFrom-failure-1-spec.k    |                     98.46 |             141.92 |                 0.6937711386696731 |
|  benchmarks/overflow00-overflow-spec.k      |                      40.5 |              58.36 |                 0.6939684715558602 |
|  benchmarks/storagevar01-spec.k             |                     60.69 |              87.41 |                 0.6943141516988903 |
|  erc20/hkg/transfer-success-2-spec.k        |                     55.03 |              79.23 |                 0.6945601413605957 |
|  erc20/ds/transferFrom-success-2-spec.k     |                     87.19 |             125.42 |                 0.6951841811513315 |
|  erc20/ds/transfer-failure-1-a-spec.k       |                     73.14 |             105.05 |                 0.6962398857686816 |
|  bihu/functional-spec.k                     |                     98.63 |             141.64 |                 0.6963428410053657 |
|  erc20/hkg/transfer-failure-2-spec.k        |                     56.08 |              80.49 |                 0.6967325133556964 |
|  erc20/hkg/transfer-failure-1-spec.k        |                     57.65 |              82.55 |                 0.6983646274984858 |
|  erc20/ds/transfer-failure-2-a-spec.k       |                     69.15 |              98.62 |                 0.7011762320016224 |
|  bihu/forwardToHotWallet-failure-3-spec.k   |                     83.66 |             117.71 |                 0.7107297595786255 |
|  bihu/forwardToHotWallet-success-1-spec.k   |                     68.55 |              96.39 |                 0.7111733582321818 |
|  benchmarks/storagevar00-spec.k             |                     43.65 |              61.33 |                 0.7117234632316973 |
|  mcd/cat-file-addr-pass-rough-spec.k        |                     60.68 |              84.35 |                 0.7193835210432721 |
|  benchmarks/requires01-a0gt0-spec.k         |                     42.77 |              59.19 |                 0.7225882750464606 |
|  bihu/forwardToHotWallet-failure-1-spec.k   |                     47.58 |              65.79 |                 0.7232102143182854 |
|  examples/storage-spec.md                   |                    271.36 |             374.76 |                 0.7240900843206319 |
|  erc20/ds/transfer-failure-2-b-spec.k       |                     49.22 |              66.44 |                 0.7408187838651414 |
|  mcd/dai-adduu-fail-rough-spec.k            |                     70.26 |              93.66 |                 0.7501601537475978 |
|  mcd/flopper-file-pass-rough-spec.k         |                    442.38 |             582.73 |                 0.7591508932095481 |
|  mcd/flipper-addu48u48-fail-rough-spec.k    |                     73.79 |              83.94 |                 0.8790802954491305 |
|  examples/solidity-code-spec.md             |                    335.66 |             365.21 |                 0.9190876482024042 |
|  mcd/vat-muluu-pass-spec.k                  |                    168.19 |             175.06 |                 0.9607563121215583 |
|  mcd/vat-mului-pass-spec.k                  |                    165.92 |             158.64 |                 1.0458900655572365 |
| TOTAL                                                                                          |         9785.380000000003 | 17274.559999999994 |                 0.5664618954115188 |